test(integration): move sample_data fixture to conftest.py
The fixture was defined identically in test_ols.py, test_tls.py,
test_comparison.py, and test_base.py. A single definition in
conftest.py is the canonical pytest location for shared fixtures and
eliminates the risk of the four copies diverging silently.
